remove some warinings

+++ src/Mesh.cpp
@@ -22,7 +22,7 @@ Mesh::Mesh(const std::vector<Vertex> &vertices, [[maybe_unused]] const std::vect
         glBindBuffer(GL_ELEMENT_ARRAY_BUFFER, EBO);
         glBufferData(GL_ELEMENT_ARRAY_BUFFER, indices.size() * sizeof(unsigned int), indices.data(), GL_STATIC_DRAW);
 #endif
-    size_t index = 0;
+    GLuint index = 0;
     // vertex positions
     glEnableVertexAttribArray(index);
     glVertexAttribPointer(index++, decltype(VetexData::Position)::length(), GL_FLOAT, GL_FALSE, sizeof(VetexData), reinterpret_cast<void*>(offsetof(VetexData, Position)) );
@@ -33,7 +33,7 @@ Mesh::Mesh(const std::vector<Vertex> &vertices, [[maybe_unused]] const std::vect
     glEnableVertexAttribArray(index);
     glVertexAttribPointer(index++, decltype(VetexData::TexCoords)::length(), GL_FLOAT, GL_FALSE, sizeof(VetexData), reinterpret_cast<void*>(offsetof(VetexData, TexCoords)) );
 
-    vInSize = vertices.size();
+    vInSize = static_cast<GLuint>(vertices.size());
 }
 
 Mesh::~Mesh()
